Amenities page design consistency fix

+++ src/service/AmenityService.java
@@ -8,7 +8,11 @@ package service;
 
 import repository.AmenityRepository;
 import beans.Amenity;
+import beans.User;
+import beans.enums.UserType;
 import exceptions.DatabaseException;
+import exceptions.InvalidUserException;
+
 import java.util.*;
 
 public class AmenityService {
@@ -21,18 +25,33 @@ public class AmenityService {
    }
 
 //Methods
-   public Amenity create(Amenity amenity) throws DatabaseException {
-	  
-	  return amenityRepository.create(amenity);
-
+   public Amenity create(String amenityName, User user) throws DatabaseException, InvalidUserException {
+	  if(user.getUserType() == UserType.admin) {
+		  Amenity amenity = new Amenity(amenityName, false);
+		  amenity = amenityRepository.create(amenity);
+		  return amenity;
+	  }
+	  else {
+		  throw new InvalidUserException();
+	  }
    }
    
-   public void update(Amenity amenity) throws DatabaseException {
-	   amenityRepository.update(amenity); 
+   public void update(Amenity amenity, User user) throws DatabaseException, InvalidUserException {
+	   if(user.getUserType() == UserType.admin) {		   
+		   amenityRepository.update(amenity);
+	   }
+	   else {
+		   throw new InvalidUserException();
+	   }
    }
    
-   public void delete(Amenity amenity) throws DatabaseException {
-	   amenityRepository.delete(amenity.getId());
+   public void delete(Amenity amenity, User user) throws DatabaseException, InvalidUserException {
+	   if(user.getUserType() == UserType.admin) {
+		   amenityRepository.delete(amenity.getId());		   
+	   }
+	   else {
+		   throw new InvalidUserException();
+	   }
    }
    
    public List<Amenity> getAll() throws DatabaseException {
